+++ src/manager/polling/poll_action.rs
@@ -1,0 +1,6 @@
+use super::PollType;
+
+#[derive(Debug, Clone)]
+pub enum PollAction {
+    DoPoll(PollType),
+}